requirements.txt
alembic==0.8.7
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.1.0
black==19.10b0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
ecdsa==0.14.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.0.4
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pathspec==0.6.0
pluggy==0.13.1
py==1.8.0
pycparser==2.14
pycryptodome==3.6.6
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
python-editor==1.0.1
python-jose-cryptodome==1.3.2
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
alembic==0.8.7
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.1.0
black==19.10b0
blinker==1.4
cachetools==4.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
ecdsa==0.14.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.2
httplib2==0.18.0
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.0.4
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.4
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
pathspec==0.6.0
pluggy==0.13.1
protobuf==3.11.3
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.14
pycryptodome==3.6.6
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
python-editor==1.0.1
python-jose-cryptodome==1.3.2
pytz==2019.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.8
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==3.1.0
psycopg2
boto3==1.12.39
